requirements.txt
# Photfolio Dependencies

# Image processing
Pillow>=10.0.0

# Template engine
Jinja2>=3.1.0

# Configuration file parsing
PyYAML>=6.0
